Using regular expressions to compare tags. We are now adding line-start and line-end characters around the pattern, and splatting stars (*) into dot-start (.*) so that the regex will act similar to how one would expect from a command line tool.
It is not likely that someone would pass something like
-tag role=^(app|cache)-[0-9]+$ on the command line. More common patterns would
be things like -tag role=*web , or -tag active=* (to match any with the tag).

Furthermore, the line-start and line-end injection is necessary, as one would
not expect something like -tag role=a to include nodes with a tag of role=cache.
In our previous implementation, this would still match due to the open-endedness
of the regular expression, and the MatchString() method. If a user really wants
to match all nodes with a tag containing the letter a, they would more likely do
something like -tag role=*a*