Fix Serf.Stats() panic when coordinates are disabled
If the Serf agent is configured with coordinates disabled (i.e.
serf.Config.DisabledCoordinates = true), invoking Serf.Stats() causes a
segmentation fault. That's because in order to set the value of
"coordinate_resets" in the stats map, Serf.Stats() invokes
s.coordClient.Stats(), with s.coordClient being a nil pointer.

This commit fixes this issue by setting the value of "coordinate_resets"
in the stats map only if coordinates are enabled.