Prevents log spam from zero RTTs.
These aren't harmful to the Vivaldi algorithm since we have a min
1 us RTT we use when this happens, so it's not worth logging them.
We added a counter so we can still observe these, since we don't
have a great root cause story on why we see them, other than a guess
that some systems have a low monotonic clock granularity.

We also added a new metric for rejected coordinates and moved the
detailed message to the TRACE level.