Fix the member validation to not always fail
The memberlist Node.Addr is an IP and guaranteed to not contain a port so the check was always going to fail. Also the Port is stored in a uint16 on the Node type so we don’t have to check that its value is within 0 - 65535 as it has to be due to the number of bytes used to store it.