Add a `MaxLeaveTime` configuration option
In suboptimal deployments, such as Consul on ECS where a Consul client is run
as a sidecar next to each service, crash looping results in inconsistencies
between serf, memberlist and raft. This change allows operators to configure
the amount of time Serf tries to gracefully shut down. On ECS, the maximum time
between a SIGTERM and SIGKILL is 120 seconds so setting `MaxLeaveTime` to
around 120 will give the leave much more time to propogate.