Rewrite paths to be internal to repo
As-per [this
ticket](https://app.asana.com/0/1207142155269470/1207142155269473/f). The
current content of these files are (were?) served up by the ruby application in
`/website` and could be accessed using `/docs/filename.html`. Since this is
being retired, this content now needs to be accessible (and navigable) from the
GitHub repository alone.

In order to accomplish this, links to `file.html` need to be rewritten
to `file.html.markdown`. The suggestion in the ticket was to use a
codemod to do this - however this could instead be accomplished using
the following `sed` expression:

```bash
find . -type f -name '*.*' -exec sed -i -E '
  s|\/intro\/(.*)\.html\)|\/docs\/intro\/\1.html.markdown\)|g;
  s|\/docs\/(.*)\.html\)|\/docs\/\1.html.markdown\)|g' {} +; git status
```