client: make consistent use of shutdownLock to protect c.shutdown
c.shutdownLock protects access to c.shutdown, but before this
commit this is only used when calling Close(),
thus only protecting against parallel user calls to Close().

It was not used when checking whether the channel is shutdown
in the exported function IsClosed(), potentially returning an
incorrect value.

It was also not used in send() and listen(), leading to many
codepaths where the lock is not properly acquired:

send() is called by Monitor(), Stream(), Query(), genericRPC(),
listen() is started as a goroutine by ClientFromConfig().

To address this, turn the c.shutdownLock Mutex into a RWMutex,
and make use of it in the missing cases, enabling use of the same
RPCClient from multiple parallel goroutines.

Fixes: https://github.com/hashicorp/serf/issues/782

Signed-off-by: Claudio Fontana <claudio.fontana@gmail.com>
Signed-off-by: Claudio Fontana <cfontana@suse.de>