client: make use of shutdownLock on Read access
c.shutdownLock protects access to c.shutdown,
but before this commit this is only used when calling Close(),
thus only protecting against parallel user calls to Close(),

but it was not used when checking whether the channel is shutdown
in the exported function IsClosed(), potentially returning an
incorrect value.

It was also not used in send() and listen().

To address this, turn the c.shutdownLock Mutex into a RWMutex,
and make use of it in all these cases, enabling use of the same
RPCClient from multiple parallel goroutines.

Fixes: https://github.com/hashicorp/serf/issues/782

Signed-off-by: Claudio Fontana <cfontana@suse.de>